feat: Add unit tests for seed liquidity single bin and lfg
